requirements.txt
librosa==0.8.1
matplotlib==3.5.1
numpy==1.21.6
pandas==1.4.2
Pillow==10.1.0
scikit_learn==1.3.2
scipy==1.7.3
seaborn==0.11.2
torch==1.13.0+cu116
torchvision==0.14.0+cu116
tqdm==4.64.0
transformers==4.34.0
